examples/requirements.txt
python-dotenv
openai[realtime]
sounddevice
numpy
langsmith
langchain-openai
langgraph
langchain